rename secretcontroller to remotesecrets
